Define the renderer
The renderer allocates a framebuffer and shares it with each tracer. The
screen is split into blocks of rows that get distributed to each tracer
each time a frame is being rendered.

A naive block distribution algorithm is currently being used where the number
of rows that get assigned to each tracer depends on the speed estimate
(over base line cpu-renderer) that each tracer reports.