Use an accumulation buffer to enable progressive rendering
The renderer uses an accumulation buffer to blend contents of previous
frames with the contents of the current frame. The idea is to set the
blendWeight to 0 when the camera moves or the frame buffer is cleared
and have the renderer automatically adjust it while the camera is not
moving.