Implement BVH trees for accelerating ray-primitive intersection tests
BVH nodes are stored in an optimized GPU-friendly structure taking 32
bytes per node and represented using 2 Vec4. The XYZ coords for the 2
vectors represent the min and max node bounds. The W components specify
either the left and right node indices (if inner node) or the index of
the first leaf primitive and the primitive count (if leaf). The sign of
the W components is used for detecting whether a node is actually an
inner node or a leaf (inner nodes have positive values).

The tracer kernel uses a fixed-stack for traversing the BVH nodes.