Move accum buffer into the tracer and fix progressive rendering bugs
By moving the accumulation buffer and final image composition into the
tracer we increase the effective rendering speed at the cost of an extra
buffer R/W operation.

All tracers have access to the same accumulation buffer and output
RGBA framebuffer (backed by a image.RGBA) but only read/write the
portions of the buffers that belong to the block that the renderer
assigns to them. Consequently, there is no overlap and tracers can still
operate in parallel.

This commit also fixes an issue with the random seed generation inside
the tracer where the seed was always truncated to 0.