Extract kernel and buffer handling into integrator package
At the moment, the integrator only implements primary ray generation but
will be extended in future commits to include the rest of kernels
required for the path tracer implementation.

The decision to refactor the integrator into a separate package was made
so that we can reuse the integrator implementation across different
tracers (e.g. to implement foveated rendereing)