Cleanup and refactor opencl tracer to support user-defined pipelines
This commit includes a huge chunk of refactoring work that improves
source code organization and hopefully makes things more readable.

Thanks to the new tracer options that can be passed to the tracer init
method we can easily assemble different tracing pipelines (e.g. to swap
the integrator or to use a different tone mapping algorithm). Here is an
example:

```
err = tr.Init(
	frameW,
	frameH,
	// Stages
	opencl.ClearAccumulator(),
	opencl.PerspectiveCamera(),
	opencl.MonteCarloIntegrator(1),
	opencl.TonemapSimpleReinhard(1.5),
	opencl.DebugFrameBuffer("debug-fb.png"),
)
```

Note:
The integrator kernel is still work in progress