[cl] Implement material node sampler and fix barrier bug in integrator
We can now properly sample layered materials and evaluate blend
functions. Adding support for blend functions means that the debug code
must also select one of the material tree leafs so it does not
mis-interpret the normal texture index (contains a >= 0 value in
material nodes due to the use of a C union)

After a bit of debugging, the PT integrator now works properly on Iris
GPUs. The code is using memory barriers to sync occlusion/indirect ray
allocations. After the first bounce, the number of rays will be less
than the number of active threads. The previous version contained a
return if(globalId < *numRays) before a barrier call. On Iris hw, if the
last workgroup is not full, the barrier is never reached by all threads
due to the return causing the kernel to crash. Surprisingly, AMD hw does
not exhibit this behavior. By refactoring the code to ensure that all
threads hit the barriers the issue is now finally resolved.