Implement helper for creating shared opencl contexts
In order to be able to copy data across devices when operating in
multi-device mode they need to share the same opencl context. This
helper receives a list of devices and creates a shared context for them.

The opencl.NewTracer() call now accepts a ctx pointer as an argument. If
that is not nil then the tracer will attempt to use the supplied
context; otherwise a new standalone context will be created.