Correctly merge tracer output when using dynamic schedulers
Due to the way that schedulers work, one tracer may be responsible for a
particular frame row in one frame and a different tracer may be assigned
to it on the following frame. This causes a problem as we need to
properly aggregate the tracer outputs to ensure that our post-processing
pipeline always operates on the same number of aggregated samples
(for example, averaging them would otherwise generate inconsistent
output).

To avoid such problems, the primary tracer maintains two accumulators, a
trace accumulator which is reset on each trace request and a frame
accumulator that accumulates (via an add operation) the assigned rows
from each tracer. Post-processing pipelines operate on the primary
tracer's frame accumulator.