Make aggregate accum. kernel async to avoid concurrency bug
When doing multi-tracer rendering it is possible to get a ABORT error
inside the opencl CGO code. It looks like the opencl wrapper is not
properly re-entrant and as a result, calling multiple times cl.Finish
inside the aggregation kernels may cause the renderer to crash.

In an attempt to fix this, the aggregate accumulator kernel that gets
invoked by multiple go-routines has been made async (does not call cl.Finish
like other kernels). The cl.Finish call has been moved inside the
SyncFramebuffer call that gets only invoked on the primary tracer. This
fix seems to reduce the crashes but does not eliminate them completely;
there are still cases where the renderer simply dies with a "signal:
abort trap" error with no further stack trace.