Implement RGB IOR dispersion via the disperse operator
The disperse operator accepts a bxdf parameter and vector params for
the int and ext IORs. When the operator is first executed a random
R, G or B channel will be selected and the appropriate IOR values will
be forced in the bxdf leaf node. In addition, we track the selected
channel throughout the path's life by setting the appropriate path
flag. This allows us to always select the same IOR when exiting the
material. Depending on the selected path, the operator evaluation
will set a primary color tint mask to simulate a single wavelength
dispersion.